+++ CHANGELOG.md
@@ -2,6 +2,16 @@ Changelog
 =========
 
 ## head
+*   Add Tag support: suffix-octothorpe, infix-comma, infix-semicolon (GH-53)
+*   Remove previously deprecated NoopClient. Use a nil `*Client` Statter as a
+    replacement, if needed. Ex:
+    ```
+    var client Client
+    // A nil *Client has noop behavior, so this is safe.
+    // It will become a small overhead (just a couple function calls) noop.
+    err = client.Inc("stat1", 42, 1.0)
+    ```
+
 
 ## 4.0.0 2020-11-05
 *   Fix go.mod versioning. (GH-51,GH-52)